+++ src/deepskylog/AstronomyLibrary/AstronomyLibrary.php
@@ -13,6 +13,7 @@
 namespace deepskylog\AstronomyLibrary;
 
 use Carbon\Carbon;
+use deepskylog\AstronomyLibrary\Coordinates\GeographicalCoordinates;
 
 /**
  * The main AstronomyLibrary class.